+++ tutor/templates/apps/openedx/settings/partials/common_all.py
@@ -13,5 +13,33 @@ BULK_EMAIL_DEFAULT_FROM_EMAIL = "no-reply@" + ENV_TOKENS["LMS_BASE"]
 API_ACCESS_MANAGER_EMAIL = ENV_TOKENS["CONTACT_EMAIL"]
 API_ACCESS_FROM_EMAIL = ENV_TOKENS["CONTACT_EMAIL"]
 
+# Load module store settings from config files
+update_module_store_settings(MODULESTORE, doc_store_settings=DOC_STORE_CONFIG)
+
+# Set uploaded media file path
+MEDIA_ROOT = "/openedx/media/"
+
+# Video settings
+VIDEO_IMAGE_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
+VIDEO_TRANSCRIPTS_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
+
+# Change syslog-based loggers which don't work inside docker containers
+LOGGING["handlers"]["local"] = {
+    "class": "logging.handlers.WatchedFileHandler",
+    "filename": os.path.join(LOG_DIR, "all.log"),
+    "formatter": "standard",
+}
+LOGGING["handlers"]["tracking"] = {
+    "level": "DEBUG",
+    "class": "logging.handlers.WatchedFileHandler",
+    "filename": os.path.join(LOG_DIR, "tracking.log"),
+    "formatter": "standard",
+}
+LOGGING["loggers"]["tracking"]["handlers"] = ["console", "local", "tracking"]
+
+EMAIL_USE_SSL = {{ SMTP_USE_SSL }}
+
+LOCALE_PATHS.append("/openedx/locale")
+
 {{ patch("openedx-common-settings") }}
 ######## End of settings common to LMS and CMS