+++ tutor/templates/apps/openedx/settings/partials/common_cms.py
@@ -2,33 +2,6 @@
 
 ######## Common CMS settings
 
-# Load module store settings from config files
-update_module_store_settings(MODULESTORE, doc_store_settings=DOC_STORE_CONFIG)
-
-# Set uploaded media file path
-MEDIA_ROOT = "/openedx/media/"
-
-# Video settings
-VIDEO_IMAGE_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
-VIDEO_TRANSCRIPTS_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
-
-# Change syslog-based loggers which don't work inside docker containers
-LOGGING["handlers"]["local"] = {
-    "class": "logging.handlers.WatchedFileHandler",
-    "filename": os.path.join(LOG_DIR, "all.log"),
-    "formatter": "standard",
-}
-LOGGING["handlers"]["tracking"] = {
-    "level": "DEBUG",
-    "class": "logging.handlers.WatchedFileHandler",
-    "filename": os.path.join(LOG_DIR, "tracking.log"),
-    "formatter": "standard",
-}
-LOGGING["loggers"]["tracking"]["handlers"] = ["console", "local", "tracking"]
-
-EMAIL_USE_SSL = {{ SMTP_USE_SSL }}
-
-LOCALE_PATHS.append("/openedx/locale")
 
 STUDIO_NAME = "{{ PLATFORM_NAME }} - Studio"