+++ tutor/templates/apps/openedx/settings/partials/common_lms.py
@@ -2,32 +2,6 @@
 
 ######## Common LMS settings
 
-# Load module store settings from config files
-update_module_store_settings(MODULESTORE, doc_store_settings=DOC_STORE_CONFIG)
-
-# Set uploaded media file path
-MEDIA_ROOT = "/openedx/media/"
-
-# Video settings
-VIDEO_IMAGE_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
-VIDEO_TRANSCRIPTS_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
-
-# Change syslog-based loggers which don't work inside docker containers
-LOGGING["handlers"]["local"] = {
-    "class": "logging.handlers.WatchedFileHandler",
-    "filename": os.path.join(LOG_DIR, "all.log"),
-    "formatter": "standard",
-}
-LOGGING["handlers"]["tracking"] = {
-    "level": "DEBUG",
-    "class": "logging.handlers.WatchedFileHandler",
-    "filename": os.path.join(LOG_DIR, "tracking.log"),
-    "formatter": "standard",
-}
-LOGGING["loggers"]["tracking"]["handlers"] = ["console", "local", "tracking"]
-
-EMAIL_USE_SSL = {{ SMTP_USE_SSL }}
-
 # Fix media files paths
 VIDEO_IMAGE_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
 VIDEO_TRANSCRIPTS_SETTINGS["STORAGE_KWARGS"]["location"] = MEDIA_ROOT
@@ -47,7 +21,6 @@ GRADES_DOWNLOAD = {
     },
 }
 
-LOCALE_PATHS.append("/openedx/locale")
 
 # JWT is authentication for other openedx services
 JWT_AUTH["JWT_ISSUER"] = "{{ JWT_COMMON_ISSUER }}"